Add currency column to Amountable models (#1362)
* Add currency columns
* Add LocalizationHelper, use default_currency helper
