+++ app/views/payments/_form.html.haml
@@ -41,17 +41,16 @@
   - payment_type = @is_payment ? 'Pay' : 'Donate'
   - description = @is_payment ? t(".payment_to_bike_index") : t(".donate_to_bike_index")
 
-  - currency = t(I18n.locale, scope: %i[money currencies])
   - stripe_form_hash = { id: 'stripe_form',
     data: { key: ENV['STRIPE_PUBLISHABLE_KEY'],
     email: current_user&.email,
     type: payment_type,
-    currency: currency,
+    currency: default_currency,
     description: description } }
 
   = form_tag payments_path, stripe_form_hash do
     = hidden_field_tag :stripe_token
     = hidden_field_tag :stripe_amount
-    = hidden_field_tag :stripe_currency, currency
+    = hidden_field_tag :stripe_currency, default_currency
     = hidden_field_tag :stripe_email
     = hidden_field_tag :is_payment, @is_payment
