+++ db/migrate/20191022143755_add_currency_to_amountable_tables.rb
@@ -1,0 +1,7 @@
+class AddCurrencyToAmountableTables < ActiveRecord::Migration
+  def change
+    add_column :theft_alert_plans, :currency, :string, null: false, default: "USD"
+    add_column :invoices, :currency, :string, null: false, default: "USD"
+    add_column :paid_features, :currency, :string, null: false, default: "USD"
+  end
+end
